+++ object/init.go
@@ -112,7 +112,7 @@ func initBuiltInUser() {
 		Id:                util.GenerateId(),
 		Type:              "normal-user",
 		Password:          "123",
-		DisplayName:       "Admin",
+		DisplayName:       "Username",
 		Avatar:            fmt.Sprintf("%s/img/casbin.svg", conf.GetConfigString("staticBaseUrl")),
 		Email:             "admin@example.com",
 		Phone:             "12345678910",
@@ -219,8 +219,8 @@ func initBuiltInLdap() {
 		ServerName: "BuildIn LDAP Server",
 		Host:       "example.com",
 		Port:       389,
-		Admin:      "cn=buildin,dc=example,dc=com",
-		Passwd:     "123",
+		Username:   "cn=buildin,dc=example,dc=com",
+		Password:   "123",
 		BaseDn:     "ou=BuildIn,dc=example,dc=com",
 		AutoSync:   0,
 		LastSync:   "",
