+++ object/ldap_conn.go
@@ -15,6 +15,7 @@
 package object
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 
@@ -23,6 +24,47 @@ import (
 	goldap "github.com/go-ldap/ldap/v3"
 	"github.com/thanhpk/randstr"
 )
+
+type LdapConn struct {
+	Conn *goldap.Conn
+	IsAD bool
+}
+
+//type ldapGroup struct {
+//	GidNumber string
+//	Cn        string
+//}
+
+type ldapUser struct {
+	UidNumber string
+	Uid       string
+	Cn        string
+	GidNumber string
+	// Gcn                   string
+	Uuid                  string
+	DisplayName           string
+	Mail                  string
+	Email                 string
+	EmailAddress          string
+	TelephoneNumber       string
+	Mobile                string
+	MobileTelephoneNumber string
+	RegisteredAddress     string
+	PostalAddress         string
+}
+
+type LdapRespUser struct {
+	UidNumber string `json:"uidNumber"`
+	Uid       string `json:"uid"`
+	Cn        string `json:"cn"`
+	GroupId   string `json:"groupId"`
+	// GroupName string `json:"groupName"`
+	Uuid        string `json:"uuid"`
+	DisplayName string `json:"displayName"`
+	Email       string `json:"email"`
+	Phone       string `json:"phone"`
+	Address     string `json:"address"`
+}
 
 func (ldap *Ldap) GetLdapConn() (c *LdapConn, err error) {
 	var conn *goldap.Conn
@@ -36,7 +78,7 @@ func (ldap *Ldap) GetLdapConn() (c *LdapConn, err error) {
 		return nil, err
 	}
 
-	err = conn.Bind(ldap.Admin, ldap.Passwd)
+	err = conn.Bind(ldap.Username, ldap.Password)
 	if err != nil {
 		return nil, err
 	}
@@ -94,6 +136,111 @@ func isMicrosoftAD(Conn *goldap.Conn) (bool, error) {
 	return isMicrosoft, err
 }
 
+func (l *LdapConn) GetLdapUsers(ldapServer *Ldap) ([]ldapUser, error) {
+	SearchAttributes := []string{
+		"uidNumber", "cn", "sn", "gidNumber", "entryUUID", "displayName", "mail", "email",
+		"emailAddress", "telephoneNumber", "mobile", "mobileTelephoneNumber", "registeredAddress", "postalAddress",
+	}
+	if l.IsAD {
+		SearchAttributes = append(SearchAttributes, "sAMAccountName")
+	} else {
+		SearchAttributes = append(SearchAttributes, "uid")
+	}
+
+	searchReq := goldap.NewSearchRequest(ldapServer.BaseDn, goldap.ScopeWholeSubtree, goldap.NeverDerefAliases,
+		0, 0, false,
+		ldapServer.Filter, SearchAttributes, nil)
+	searchResult, err := l.Conn.SearchWithPaging(searchReq, 100)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(searchResult.Entries) == 0 {
+		return nil, errors.New("no result")
+	}
+
+	var ldapUsers []ldapUser
+	for _, entry := range searchResult.Entries {
+		var user ldapUser
+		for _, attribute := range entry.Attributes {
+			switch attribute.Name {
+			case "uidNumber":
+				user.UidNumber = attribute.Values[0]
+			case "uid":
+				user.Uid = attribute.Values[0]
+			case "sAMAccountName":
+				user.Uid = attribute.Values[0]
+			case "cn":
+				user.Cn = attribute.Values[0]
+			case "gidNumber":
+				user.GidNumber = attribute.Values[0]
+			case "entryUUID":
+				user.Uuid = attribute.Values[0]
+			case "objectGUID":
+				user.Uuid = attribute.Values[0]
+			case "displayName":
+				user.DisplayName = attribute.Values[0]
+			case "mail":
+				user.Mail = attribute.Values[0]
+			case "email":
+				user.Email = attribute.Values[0]
+			case "emailAddress":
+				user.EmailAddress = attribute.Values[0]
+			case "telephoneNumber":
+				user.TelephoneNumber = attribute.Values[0]
+			case "mobile":
+				user.Mobile = attribute.Values[0]
+			case "mobileTelephoneNumber":
+				user.MobileTelephoneNumber = attribute.Values[0]
+			case "registeredAddress":
+				user.RegisteredAddress = attribute.Values[0]
+			case "postalAddress":
+				user.PostalAddress = attribute.Values[0]
+			}
+		}
+		ldapUsers = append(ldapUsers, user)
+	}
+
+	return ldapUsers, nil
+}
+
+// FIXME: The Base DN does not necessarily contain the Group
+//
+//	func (l *ldapConn) GetLdapGroups(baseDn string) (map[string]ldapGroup, error) {
+//		SearchFilter := "(objectClass=posixGroup)"
+//		SearchAttributes := []string{"cn", "gidNumber"}
+//		groupMap := make(map[string]ldapGroup)
+//
+//		searchReq := goldap.NewSearchRequest(baseDn,
+//			goldap.ScopeWholeSubtree, goldap.NeverDerefAliases, 0, 0, false,
+//			SearchFilter, SearchAttributes, nil)
+//		searchResult, err := l.Conn.Search(searchReq)
+//		if err != nil {
+//			return nil, err
+//		}
+//
+//		if len(searchResult.Entries) == 0 {
+//			return nil, errors.New("no result")
+//		}
+//
+//		for _, entry := range searchResult.Entries {
+//			var ldapGroupItem ldapGroup
+//			for _, attribute := range entry.Attributes {
+//				switch attribute.Name {
+//				case "gidNumber":
+//					ldapGroupItem.GidNumber = attribute.Values[0]
+//					break
+//				case "cn":
+//					ldapGroupItem.Cn = attribute.Values[0]
+//					break
+//				}
+//			}
+//			groupMap[ldapGroupItem.GidNumber] = ldapGroupItem
+//		}
+//
+//		return groupMap, nil
+//	}
+
 func LdapUsersToLdapRespUsers(users []ldapUser) []LdapRespUser {
 	res := make([]LdapRespUser, 0)
 	for _, user := range users {
@@ -135,7 +282,7 @@ func SyncLdapUsers(owner string, respUsers []LdapRespUser, ldapId string) (*[]Ld
 	affiliation := strings.Join(dc, ".")
 
 	var ou []string
-	for _, admin := range strings.Split(ldap.Admin, ",") {
+	for _, admin := range strings.Split(ldap.Username, ",") {
 		if strings.Contains(admin, "ou=") {
 			ou = append(ou, admin[3:])
 		}
