+++ web/src/table/LdapTable.js
@@ -45,8 +45,8 @@ class LdapTable extends React.Component {
       serverName: "Example LDAP Server",
       host: "example.com",
       port: 389,
-      admin: "cn=admin,dc=example,dc=com",
-      passwd: "123",
+      username: "cn=admin,dc=example,dc=com",
+      password: "123",
       baseDn: "ou=People,dc=example,dc=com",
       autosync: 0,
       lastSync: "",
